refactor(chain_manager): move storable tests to separate file
